refactor move spec test to test folder
