+++ jest.config.js
@@ -3,5 +3,5 @@
 module.exports = {
   preset: "ts-jest",
   testEnvironment: "node",
-  roots: ["src"],
+  roots: ["test"],
 };
